Keep exception handler typing tests with application tests
